IPC doing

+++ inc/env.h
@@ -67,8 +67,8 @@ struct Env {
 	envid_t env_ipc_from;		// envid of the sender
 	int env_ipc_perm;		// Perm of page mapping received
 
-	// for priority scheduling
-	uint32_t env_priority;
+	// for priority scheduling (level 1-32)
+	uint8_t env_priority;
 };
 
 #endif // !JOS_INC_ENV_H
