+++ kern/env.h
@@ -12,9 +12,9 @@ extern struct Segdesc gdt[];
 
 void	env_init(void);
 void	env_init_percpu(void);
-int	env_alloc(struct Env **e, envid_t parent_id, uint32_t priority);
+int	env_alloc(struct Env **e, envid_t parent_id, uint8_t priority);
 void	env_free(struct Env *e);
-void	env_create(uint8_t *binary, enum EnvType type, uint32_t priority);
+void	env_create(uint8_t *binary, enum EnvType type, uint8_t priority);
 void	env_destroy(struct Env *e);	// Does not return if e == curenv
 
 int	envid2env(envid_t envid, struct Env **env_store, bool checkperm);
